Update gitweb to support clone over HTTPS + more
Removes the git-static subdomain – no longer needed. Also adds a systemd
timer and service to update the git server info in the repositories,
which is needed for fetch over dumb HTTP.

Also restructures some Apache config into a more modular form, using
Include. I want to move more of my Apache config to that model, but
that’ll have to wait for a bit.