Specify utf-8 in gitweb.conf
This doesn’t fix the problem I had, but let’s do it anyways.

(The problem I had was that author names were being served in latin1
(Felix Dörre). My fix is to apply s/D.rre/Dörre/g to $author and
$displaytext in format_search_author, /usr/share/gitweb/gitweb.cgi,
which is ugly as sin, but works.