Add dehydrated config
dehydrated (formerly known as letsencrypt.sh) is a simple ACME client.
This commit adds the default Debian configuration for it, as well as
some additional configuration to make it work under apache2 and to run
it as a systemd service (with associated timer).

The service runs as the dehydrated user, and stores its persistent data
(accounts, certificates) in /var/lib/dehydrated and its runtime data
(ACME challenges) in /run/dehydrated. With the configuration files added
here, the following commands can be used to set up the system for it:

    systemd-sysusers dehydrated.conf
    systemd-tmpfiles --create dehydrated.conf

Most of the sandboxing options in dehydrated.service are not necessary
for a non-root service; they are present because I added them before I
changed the config to run dehydrated as non-root, and because they don’t
hurt to have either.