Add SpamAssassin configuration
Debian ships services for SpamAssassin, but they’re so weak (short unit
file for spamassassin, unmigrated SysV service for spamass-milter) that
I’d rather start over with my own unit files.

The sandbox is copied from dehydrated.service (with the ReadWritePaths
tweaked), and so it still contains settings that aren’t necessary for a
non-root service, but don’t hurt to have either.

The spamass-milter setup differs slightly from the Ars article [1]:
instead of chmoding and chowning the socket after spamass-milter
launches (which the init script does), we add postfix to the
spamass-milter group and make the socket group-accessible by tweaking
the umask. This is analogous to the opendkim and opendmarc setup, down
to the umask trick (which for those services is found in their
respective configuration files). (You can also do the chmod+chown in
ExecStartPost, but then you can’t use PrivateUsers=yes, since that hides
the postfix group even from commands prefixed with ‘+’.)

[1]: https://arstechnica.com/business/2014/03/taking-e-mail-back-part-3-fortifying-your-box-against-spammers/