Rename 000-default.conf to zzz-default.conf
If the symlink is called 000-default.conf, it breaks lucaswerkmeister.de
(not sure if just over HTTP or also over HTTPS). The old symlink stored
in git was already named zzz-default.conf, but if the original file
stays named 000-default.conf, it’s susceptible to a2ensite or default
configuration.