Add sandboxing directives for apache2.service
I put this off for a long time because it was scary (lots of stuff in
Apache!), and also (back on the Jessie server) because apache2 was still
a SysV service back then. But actually this wasn’t too hard to put
together and seems to work. (Famous last words…)