Mark git-l emails that don’t involve me as read
The git mailing list is pretty high-volume, so let’s mark emails that
don’t mention me (To or Cc) as seen right off the bat. I can still look
over them if I feel like it, but I can also not look at the folder for a
few days or weeks and it won’t grow to contain thousands of unread
emails.

(The “involves me” part is not tested yet, so for the next few days I’ll
try to keep a lookout, just in case it’s broken.)

I would’ve liked to keep this in read.sieve, where it makes more sense,
but it doesn’t seem to work. My current guess is that performing any of
the four Actions (keep, fileinto, redirect, discard) terminates
processing of a message, and since read.sieve runs after
mailing-lists.sieve (r > m), the fileinto in mailing-lists.sieve means
that read.sieve never runs for these emails. Perhaps I’ll restructure
this in the future into something like 10-read.sieve,
90-mailing-lists.sieve – not sure.