Improve InaccessiblePaths in sandboxes
Reorder the paths by decreasing likeliness to exist on an unknown system
(estimated, not scientific – this is just for readability), and prefix
all the paths with `-` so that systemd doesn’t report an error if a path
don’t exist.

This should make it possible to copy the option to other machines more
easily. (Without the `-`, you would get errors e. g. on a system without
OpenDKIM installed.)