Add tmpfiles.d config for apache2 lock directory
The apache2.service unit file specifies /var/lock/apache2 as a writeable
directory, which means that the unit fails to start if the directory
does not exist. Let’s make sure it exists via systemd-tmpfiles.

(I’m not sure if this should actually be owned by www-data or by root…)