Rename gitweb.conf to gitweb-proxy.conf
The postinst scripts of the gitweb package automatically enable a
“gitweb” configuration (i. e., symlink it from conf-available into
conf-enabled). This gitweb.conf very much is not intended to be included
unconditionally, and the package doing so has broken my website several
times. Since we probably can’t get the gitweb package to quit this
rubbish, rename the configuration instead.