Add sandbox for irqbalance.service
I straced irqbalance to find out how it works, and it turns out it
doesn’t need any special capabilities or system calls, it just reads and
writes files in /proc/irq. This means that we should be able to apply a
fairly heavy sandbox to it. (The daemon already drops all its
capabilities when started – commendable!)