Reload Apache after renewing certificates
After dehydrated is done renewing the site certificates, reload apache2
to ensure the new certificates are used.

It would be nicer to do this only if necessary (most of the time the
certificates were not changed), but this should be fine for now.

Note that `systemctl reload apache2` translates to `apachectl graceful`,
which performs a full, though graceful, restart of Apache, including
re-initializing the certificates.