Add redirects for my bachelor’s thesis
I originally dumped my bachelor’s thesis in /var/www as thesis.pdf, but
now that I’m working on my master’s thesis [1], that’s ambiguous. Issue
a permanent redirect from /thesis.pdf to /bachelor-thesis.pdf, the new
location of that file (and also for its -signed version), as long as the
master’s thesis is not done yet; once that is online as well, I’ll
switch the redirect to be a temporary one to /master-thesis.pdf.

[1]: https://github.com/lucaswerkmeister/master-thesis