Remove @resources from system call filters
In spamassassin.service, filtering the setresource() system call causes
Pyzor to die. But in general, I’m not sure why I decided to filter that
system call group – I see no reason to do so *for non-root services*,
and all of these services run as custom users. If they want to limit
their resource use beyond what the unit file already configures, that
should not be a problem.