Update redirects for my thesis
As announced in 74089316fc, thesis.pdf now redirects to
master-thesis.pdf rather than bachelor-thesis.pdf, and the same applies
to the -signed versions. According to the file modification time, I
updated this on 15 October 2018, 11:45 CEST, but I forgot to commit it
until I had to make an unrelated commit just now (October 31).