Restart Tor every twelve hours
There appears to be some memory leak: about once a day (the interval is
somewhere around 26 hours), the kernel will run out of memory and kill a
process, usually Tor. And though the memory occupied by Tor does not
appear to be anywhere near the memory limit of my server (neither
according to htop – i. e., the individual process’ virtual, resident or
shared memory – nor according to memory accounting on the service’
cgroup), restarting it (whether through the OOM killer or manually)
reliably resolves the out-of-memory situation. I suspect the leak is
somewhere in the accounting logic (if there’s a leak for each network
packet counted in the cgroup, that would accumulate quickly), though I’m
not sure if it’s in the kernel or in systemd.

Either way… I don’t have a good way to investigate this further, so for
now let’s paper over the problem by restarting Tor every twelve hours.
(Yes, I’m aware I’m applying xkcd #1495 [1] here.) I’ll try to remember
to disable the restart at some point in the future and check if it’s
still necessary or not.

[1]: https://xkcd.com/1495/