Fix SpamAssassin configuration for Debian
Debian wants to call the SpamAssassin user debian-spamd instead of
spamd. Convincing it to use our name requires several steps, only some
of which can be configured in /etc, unfortunately, so the others are
described in this commit message.

dpkg includes a system to override the default status (owner, group,
mode) of paths, dpkg-statoverride. Package install scripts can query
this to find out whether they should adjust the status or not.
Basically, it’s a rudimentary version of systemd-tmpfiles(8) /
tmpfiles.d(5), but with only a single opaque configuration file with no
comment support that’s not stored below /etc, and which requires
explicit cooperation from all install scripts and is Debian-specific.

To inform this system that we would like to have /var/lib/spamassassin
be owned by spamd, run the following command:

    dpkg-statoverride --update --add spamd spamd 0755 /var/lib/spamassassin

Unfortunately, the spamassassin.postinst script queries the database for
/var/lib/spamassassin/* instead of /var/lib/spamassassin, so we also
have to add some path below that directory to the database. I chose
sa-update-keys, but I imagine any other would also work, as long as it’s
guaranteed to exist across version updates. (Note the different mode!)

    dpkg-statoverride --update --add spamd spamd 0700 /var/lib/spamassassin/sa-update-keys

The spamassassin.postinst script then attempts to run a command as this
user, without specifying a shell, so we have to change spamd’s shell
from /sbin/nologin to /bin/sh. This commit does that in the
sysusers.d(5) configuration file, but re-running systemd-sysusers(8)
will not update the configuration of existing users, so the shell also
has to be updated manually:

    chsh -s /bin/sh spamd

Debian has no own mechanism like sysusers.d, as far as I can tell, so
there’s no way to prevent the creation of an entirely superfluous
debian-spamd user, unfortunately. (More specifically, I can see that the
creation of that user by its hard-coded name in spamassassin.postinst is
not guarded by any condition other than it already existing.)

With all these changes, the following finally succeeds:

    dpkg --configure spamassassin