Junk emails that pretend to be from me but aren’t
If an email pretends to be from my domain in the message header, but is
actually from somewhere else on the SMTP level, move it into junk. This
filters out these “account compromised” spam emails (some of which made
it past my regular spam filter).

This can be tested manually:

    $ telnet lucaswerkmeister.de 587
    Trying 2a01:4f8:c0c:2613::2...
    Connected to lucaswerkmeister.de.
    Escape character is '^]'.
    220 lucaswerkmeister.de ESMTP
    HELO lucaswerkmeister.de
    250 lucaswerkmeister.de
    MAIL FROM:<nobody@example.com>
    250 2.1.0 Ok
    RCPT TO:<mail@lucaswerkmeister.de>
    250 2.1.5 Ok
    DATA
    354 End data with <CR><LF>.<CR><LF>
    From: Lucas Werkmeister <mail@lucaswerkmeister.de>
    To: Lucas Werkmeister <mail@lucaswerkmeister.de>
    Subject: Test

    Halli hallo.
    .
    250 2.0.0 Ok: queued as [masked]
    QUIT
    221 2.0.0 Bye
    Connection closed by foreign host.