Update SpamAssassin user for Buster
Beyond all the changes laid out in c806a8f2a7, for the version in Buster
we apparently need one more: the home directory of the spamd user must
be the spamassassin data directory (otherwise sa-compile will try to
cache rules somewhere in //.spamassassin/sa-compile.cache).

As with the default shell, applying this change to sysusers.d won’t
automatically apply it to a system where the spamd user already exists,
so I also had to run:

    systemctl stop spamassassin
    usermod -d /var/lib/spamassassin spamd
    systemctl start spamassassin

(The service needs to be stopped in between because usermod refuses to
modify a user who has processes running.)

Also, the postinst script at /var/lib/dpkg/info/sa-compile.postinst
tries to chmod files in the cache directory as the hard-coded
debian-spamd user even though just a few lines above it has determined
the correct user to use and stored it in the $OWNER variable. I seem to
recall having this problem before, but I didn’t record a solution in the
commit message of c806a8f2a7; anyways, for now my solution is to edit
the script, replacing debian-spamd with $OWNER. (Unquoted, because the
rest of the script doesn’t do correct quoting either, so why bother.)

I also had one file that the script couldn’t chown because for some
reason it was owned by root; that might just be spurious, but anyways,
my fix for this was simple:

    unlink /var/lib/spamassassin/compiled/5.024/3.004002/auto/Mail/SpamAssassin/CompiledRegexps/body_0/body_0.so