Run Minecraft server all the time
With several interested players now on the whitelist, it doesn’t really
make sense to only run it when I want to play. Remove the conflict with
Tor – I think the two services can actually run at the same time, it’s
Superperm that’s problematic. But I’ll check the log for “can’t keep up,
is the server overloaded?” messages later.