Remove IP addresses from Minecraft logs
The Minecraft server logs the IP addresses of everyone who connects, but
I don’t want those – IP addresses are personal data, GDPR, yadda yadda.
There doesn’t seem to be an option to disable this behavior, so my
solution is to pipe the output through a sed filter that looks for IPv4
and IPv6 addresses (for IPv6, only in approximation, to avoid matching
non-IP addresses) and replaces them with placeholders.

This also means we should now set the SyslogIdentifier, otherwise it’ll
be “sh”. (Really, I should’ve done that before as well, since it was
“java” rather than “minecraft”. Oh well.)

The --unbuffered option for sed tries to ensure that as little delay in
the log timestamps as possible is introduced.