Trim leftover shell process from Minecraft service
sh -c 'a | b' leaves three processes running: a, b, and the surrounding
shell. I’m not aware of any standard/POSIX way to avoid this, but with a
Bashism there’s a way: bash -c 'exec a > >(exec b)' spawns b and then
replaces the shell with a (connecting them with an equivalent pipe).
bash -c 'exec b < <(exec a)' is functionally equivalent, but puts the
focus on b, which in this case makes less sense.