Add more sandboxing options to gitweb.service
Mainly those suggested by `systemd-analyze security gitweb.service`.
(Also, rearrange ProtectHome= so the list is in alphabetical order –
though I guess that the previous almost-alphabetical order was not a
deliberate choice at the time, but rather the result of systemd.exec(5)
previously ordering the optional alphabetically.)

One setting we notably cannot use is DynamicUser= – we need to set a
static Group=, and if I read the manpage correctly, it’s not possible to
combine a dynamic user with a static group or vice versa.