Create and use static gitweb user
As mentioned in the parent commit message, we can’t use DynamicUser=
because (I think) it doesn’t work well with a static group. However, we
can at least allocate a static user that’s different from the nobody
user, and use that for the service.

Also, now that the service doesn’t share its user with anything else, we
can add RemoveIPC=yes. (I didn’t want to add that while the service
still ran as User=nobody.)

Note that the user as such is not part of the www-data group, that only
applies to gitweb.service.