Update Minecraft to 1.17
This requires Java 1.16 or later, so we need a new Java. There’s no
appropriate Java version in Debian Buster (not even backports), and I
don’t want to deal with setting up testing as a repository (though
reportedly that has Java 17), but fortunately GraalVM 21.1.0 includes an
experimental Java 16 version that we can use.