Clean up /var/lib/dehydrated/certs after 120 days
Apparently, dehydrated does not clean up old certificates by itself –
which is fair enough for a “do one thing and do it well”-style tool, but
also meant that I had files in there dating back to July of 2017 (the
date of commit 37105c7ade, to be exact). Instruct systemd-tmpfiles to
clean them up after 120 days: the certificate validity period is 90
days, and dehydrated regenerates them after 60 days, so certificates
older than 120 days should be safe to remove. (Also, systemd-tmpfiles
inspects the file access date as well, so if the files *are* still read
by something, that might also prevent cleanup – though I assume that
also depends on whether the file system is mounted with atime, noatime,
relatime, or strictatime, and I haven’t tested this.)