Add additional sandbox directives for Minecraft
Lowering the `systemd-analyze security` score from 3.9 to 2.0 (though
both are OK 🙂).

This doesn’t fix what’s probably the biggest security hole – that the
server is allowed to write to its whole installation directory,
including the server JAR, not just the world – but that’ll take a bit
more work to fix, so I’ll commit it separately.