Move Minecraft world to /var/lib, add ProtectSystem=strict
The whole contents of /opt/minecraft move into /var/lib/minecraft, and
/opt/minecraft is a fresh directory, now root-owned, with only the
server JAR and nothing else. This at least prevents the Minecraft server
from overwriting its own JAR, and also lets us make the rest of the
system read-only; that said, it still leaves something to be desired. It
looks like Minecraft insists on placing various other directories in its
current working directory, including other JARS (under libraries/, and
also a copy of the server JAR under versions/), some of which probably
need to be writable at runtime, when really ideally I’d want it to write
only the world file and nothing else. But as far as I can tell, it’s not
really possible to configure it in a way that makes e.g. the ops.json or
whitelist.json unwritable.