Remove CAP_NET_BIND_SERVICE from Tor’s CapabilityBoundingSet
I assume this is in the upstream service file because Tor can be
configured to listen on privileged ports, and it’s actually not uncommon
to have it listen on ports 80 or 443, since those are less often
firewalled. However, we run Tor on non-privileged ports (ORPort 9001,
DirPort 9030, SOCKSPort 9050, ControlPort 9051), so it shouldn’t need
this capability, and seems to work fine without it in practice.