Partially sandbox cron.service
Ideally, I’d like to not even have cron installed: everything should be
a systemd timer, and each corresponding service could be sandboxed
individually. However, that doesn’t seem possible with current Debian.

This sandbox is still fairly open – cron continues to run as root, in
the host’s network namespace, with several important capabilities and
the ability to write to significant parts of the file system – but
removes tons of more exotic abilities and capabilities, as well as some
significant standard ones. Most significant is probably the
ProtectSystem=full (not strict, but at least /usr, /boot and /etc are
protected).

I’ll keep an eye out over the following days and weeks whether this
causes any apparent problems.