Limit mariadb.service to localhost
I only want mariadb to listen on localhost (it’s used for a MediaWiki
installation), so let’s enforce that using systemd.

Implemented using two sets of directives, IPAddressAllow=/IPAddressDeny=
and RestrictNetworkInterfaces=. I’m not sure if there’s a benefit to
using both, but why not, I figure. (RestrictNetworkInterfaces= was added
in systemd v250.)