Remove AF_UNIX from git@.service
This isn’t recommended in general, but we already removed /run from
git@.service (33ee59cdfa), which is where sockets to connect to ought to
live, and that’s been working fine. Let’s disconnect the Git daemon
(which gets a fair amount of dubious traffic, judging by the journal)
even further from the rest of the system.