Remove /dovecot/ from .gitignore
The only file we don’t want to commit is /passwd.db (the passwords are
hashed but there’s still no reason to let everyone crack them). There
was also a private/ directory, with dovecot.{key,pem} symlinks to
snakeoil files in /etc/ssl, but that seems to have been unused – I
renamed it a month ago and nothing broke.