Update Taavi's apt repository
I'm consolidating some of my infrastructure, apt.taavi.wtf is being
retired and merged to apt.majava.org. The taavi-keyring (now
majava-keyring) package has been updated for apt.majava.org as well, so
it should work out of the box as long as you upgrade that before
applying this patch.