Sandbox mdmonitor.service a bit
It’s a fairly loose sandbox, both because I don’t understand this
service very well (but it sounds important) and because I suspect it
needs to stay in the root mount namespace, which would mean a lot of the
sandboxing options aren’t available. I used this command to check
whether a particular directive touches the mount namespace:

    systemd-run -p ProtectHostname=yes --pipe --wait --quiet ls -l /proc/1/ns/mnt /proc/self/ns/mnt | sed 's|/1/|/   1/|'

Still, it’s better than nothing. The best thing is PrivateNetwork=yes –
I was motivated to do this by an abuse email similar to the one
described in [1], and while I believe it’s the same situation as
described there (i.e. nothing is actually sending malicious traffic from
my server), it’s still nice to have PrivateNetwork=yes on more services.

[1]: https://delroth.net/posts/spoofed-mass-scan-abuse/