Set up regular backups of my emails
After the latest mishap [1][2] I’d like to have backups of my email
folder. The simplest way is probably this – just have timers on
different frequencies copy to different directories, and for “rotation”
just delete the directory immediately before recreating it.

The sandbox needed some trial and error but is now quite tight. (Just in
case… um… there’s a vulnerability in /bin/rm or /bin/cp that can be
triggered by the contents of the files they operate on? Listen. I don’t
care.)

[1]: https://wikis.world/@LucasWerkmeister/113545141242633194
[2]: https://lucaswerkmeister.de/posts/drafts/email-bankruptcy/